removed equivalence cache tests
fix build
